requirements.txt
azure-ai-formrecognizer==3.3.3
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.29.7
azure-storage-blob==12.19.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.1.1
pillow==10.2.0
pycparser==2.22
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
